+++ docs/source/api/lr_reduction.utils.rst
@@ -23,18 +23,18 @@ lr_reduction.utils.deprecated
    :undoc-members:
    :show-inheritance:
 
-lr_reduction.utils.files
-++++++++++++++++++++++++++
-
-.. automodule:: lr_reduction.utils.files
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 lr_reduction.utils.logging
 ++++++++++++++++++++++++++++
 
 .. automodule:: lr_reduction.utils.logging
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+lr_reduction.utils.xml_to_yaml
+++++++++++++++++++++++++++++++++
+
+.. automodule:: lr_reduction.utils.xml_to_yaml
    :members:
    :undoc-members:
    :show-inheritance:
