+++ pyproject.toml
@@ -19,7 +19,7 @@ repository = "https://github.com/neutrons/LiquidsReflectometer"
 documentation = "https://lr-reduction.readthedocs.io/latest/index.html"
 
 [project.scripts]
-lr-xml-to-yaml = "lr_reduction.io.xml_to_yaml:main"
+lr-xml-to-yaml = "lr_reduction.utils.xml_to_yaml:main"
 
 #####################################
 #  BUILDING FROM SOURCES WITH HATCH
