+++ src/lr_reduction/io/__init__.py
@@ -1,11 +1,9 @@
 from lr_reduction.io.config_loader import ConfigLoader
 from lr_reduction.io.run_loader import RunLoader
-from lr_reduction.io.xml_to_yaml import convert as convert_xml_to_yaml
 from lr_reduction.models.run_data import RunData
 
 __all__ = [
     "ConfigLoader",
     "RunData",
     "RunLoader",
-    "convert_xml_to_yaml",
 ]
