+++ src/lr_reduction/io/config_loader.py
@@ -2,10 +2,10 @@ from pathlib import Path
 
 from lr_reduction.io.interfaces import ConfigLoaderInterface
 from lr_reduction.io.orso import read_config as read_orso_config
-from lr_reduction.io.xml import read_config as read_xml_config
 from lr_reduction.io.yaml import read_config as read_yaml_config
 from lr_reduction.models.config import ReductionConfig
 from lr_reduction.utils import deprecated, get_logger
+from lr_reduction.utils.xml_to_yaml import read_config as read_xml_config
 
 logger = get_logger(__name__)
