+++ src/lr_reduction/utils/__init__.py
@@ -1,9 +1,9 @@
 from lr_reduction.utils.deprecated import deprecated
-from lr_reduction.utils.files import get_sequence_id_from_path
 from lr_reduction.utils.logging import get_logger
+from lr_reduction.utils.xml_to_yaml import convert as convert_xml_to_yaml
 
 __all__ = [
-    "get_sequence_id_from_path",
     "get_logger",
     "deprecated",
+    "convert_xml_to_yaml",
 ]
